Export static lib (#25)
Making a library with AFNetworking... symbols (the ones we loaded in our dependency) makes it impossible for our clients to use AFNetworking as well as their library symbol names would collide with ours.
Starting from the last point, the only valid approach is to rename the symbols of the library we are using somehow, the most common solution is to prefix everything with our own namespace so that we use FUP_AFNetworking... instead of plain AFNetworking.... That's what I describe in the readme, the process is to extract all the symbols and create a header with macros that will do the renaming automatically. The header has to be processed before loading anything from the library, that's why there is a SDKPrefixHeader.h file and it's configured as a Prefix Header in the project.

Having solved that, there is no sane way to do the namespacing and still use cocoapods due to all the configurations and how cocoapods works in general (by not building dependencies explicitly). It used to be a hack using the Pods-environment.h header but the cocoapods creator wisely decided to remove that injection point so...

I have added the AFNetworking library dependency inside the Demo project to test that it won't have conflicts with the symbols being loaded in the SDK.

Finally, I have created a sibling repository to store the library versions so that people can download the SDK (cocoapods requires it). It is now a submodule of this repository so that it's easier to prepare SDK releases. I also created a utility script to make sure we follow the right steps to create a release of the SDK in cocoapods.

With all this, we should be able to publish a new version of the SDK in cocoapods with no problem, we will see how everything works the first time we try it, in the meantime I tried to load it locally and it works perfectly.