Add persistence (#28)
We already had some storage implementations for Config and Metric values, we were just storing them in memory. I reused those so that upper layers are unaffected.

New implementations are now depending on FUPSqlite which is just a helper class to group some common SQL logic inside (running everything in the same thread, keeping a single open connection to the database to avoid problems, creating tables for the first time lazily, etc).

I did not create a reports table yet (as we have in Android) because for now it's not needed but that will probably come in a future PR whenever we include new metrics.