Frame time metric (#34)
We had everything around CPU metrics, we changed that so that there are now only metrics. There are some parameters that are different when you compare CPU and UI metrics (CPU is about a single value taken every N seconds, frame time takes a bunch of statistical values like mean and some percentiles every N seconds). To model those differences I just used a plain old dictionary instead of going for the subclassing route, the main advantage of doing it like that is that we don't have to fight creating M subclasses and mappings, we just use JSON serialization. The drawback is that we obviously miss some type checking in the process.

Add a version number to every table so that if the schema changes we can safely drop the table and continue storing values.

Frame time is captured by using a CADisplayLink that is thought exactly for that.