Disk storage metric (#37)
Add a new collector to read all the files in Documents and Library directories and store the results in persistence, ready to be reported when the report scheduler kicks in. We are reading files size with the NSFileManager class and traversing every single file in the target directory to see how much space do they take.