Safety net (#39)
By using the @try/@catch as described in here: https://developer.apple.com/library/content/documentation/Cocoa/Conceptual/Exceptions/Tasks/HandlingExceptions.html . Errors are reported to server through its API.

We are using the safety net in every execution of the schedulers. The only code not covered by this is the main configuration in the FlowUp class.

As far as I tested we can capture everything that is represented with a NSException, including array out of bounds and more.