WiFi only reporting (#46)
Use the Reachability class found here: https://developer.apple.com/library/content/samplecode/Reachability/Introduction/Intro.html to detect when the device changes the connection from WiFi to not-WiFi.

Fix queues usage (the logic to use async/sync calls was reversed). Besides, we are now using a QueueStorage to make sure queues are singleton and we are not creating several queues for the same class (and in that way, moving the singleton state out of those classes).